refactor tests of Diffraction2D into test_diffraction2d.py
